# Speed up `iter_models` (x2-x3 improvement)