It is actually quite slow to iterate over a polars dataframe with the index, it it better to use `iter_rows`.
```python

import patito as pt

class MyModel(pt.Model):
    a: int
df = MyModel.DataFrame(pl.DataFrame({"a": range(1_000_000)}))

# main
%timeit -r1 -n1 list(df.iter_models(validate_model=False)) # 7.56 s
%timeit -r1 -n1 list(df.iter_models(validate_model=True)) # 6.62 s

# pr
%timeit -r1 -n1 list(df.iter_models(validate_model=False)) # 3.53 s
%timeit -r1 -n1 list(df.iter_models(validate_model=True)) # 2.39 s 
```
Maybe we should expose `buffer_size` as well ?

NB: Weird that it is faster to validate the model, right ?