@thomasaarholt 2024-10-22
> NB: Weird that it is faster to validate the model, right ?
Yes. I immediately checked the dataframe dtypes and height, but they are unchanged. I have no idea what is going on there 😅