@thomasaarholt 2024-10-22
Good work! Let's leave `buffer_size` out of the equation until someone needs it.